// src/modules/about/interfaces/about.interfaces.ts
export interface StaffMember {
    id: number;
    name: string;
    role: string;
    category: string;
    image: string;
}

export interface StatItem {
    label: string;
    value: string;
}

export interface Testimonial {
    name: string;
    role: string;
    text: string;
    rating: number;
}